+++ tests/FSharp.Data.Tests/CsvProvider.fs
@@ -21,18 +21,25 @@ let [<Literal>] simpleCsv = """
 type SimpleCsv = CsvProvider<simpleCsv>
 
 [<Test>]
-let ``Bool column correctly infered and accessed`` () = 
+let ``Bool column correctly inferred and accessed`` () = 
   let csv = SimpleCsv.GetSample()
   let first = csv.Rows |> Seq.head
   let actual:bool = first.Column1
   actual |> should be True
 
 [<Test>]
-let ``Decimal column correctly infered and accessed`` () = 
+let ``Decimal column correctly inferred and accessed`` () = 
   let csv = SimpleCsv.GetSample()
   let first = csv.Rows |> Seq.head
   let actual:decimal = first.Column3
   actual |> should equal 3.0M
+
+[<Test>]
+let ``Guid column correctly inferred and accessed from mislabeled TSV`` () = 
+  let csv = CsvProvider<"Data/TabSeparated.csv", HasHeaders=false>.GetSample()
+  let first = csv.Rows |> Seq.head
+  let actual:Guid option = first.Column3
+  actual |> should equal (Some (Guid.Parse("f1b1cf71-bd35-4e99-8624-24a6e15f133a")))
 
 [<Test>]
 let ``Guid column correctly infered and accessed`` () =